Revisit trait handling, shader factory to interface

+++ JavaGraphics/src/com/graphics/lib/canvas/CanvasObject.java
@@ -20,7 +20,6 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
-import com.google.common.collect.Sets;
 import com.graphics.lib.Facet;
 import com.graphics.lib.GeneralPredicates;
 import com.graphics.lib.LightIntensityFinderEnum;
@@ -35,6 +34,7 @@ import com.graphics.lib.interfaces.ICanvasObject;
 import com.graphics.lib.interfaces.ILightIntensityFinder;
 import com.graphics.lib.interfaces.ITrait;
 import com.graphics.lib.interfaces.IVertexNormalFinder;
+import com.graphics.lib.traits.TraitHandler;
 import com.graphics.lib.transform.Transform;
 
 /**
@@ -63,7 +63,7 @@ public class CanvasObject extends Observable implements ICanvasObject {
     private IVertexNormalFinder vnFinder = VertexNormalFinderEnum.DEFAULT.get();
     private Optional<WorldCoord> fixedCentre = Optional.empty();
     
-    private Set<ITrait> traits = Sets.newHashSet();
+    //private Set<ITrait> traits = Sets.newHashSet();
 	private final int objectId = nextId++;
 	
 	public CanvasObject(Supplier<Pair<ImmutableList<WorldCoord>, ImmutableList<Facet>>> initMesh) {
@@ -79,20 +79,17 @@ public class CanvasObject extends Observable implements ICanvasObject {
     }
 	
 	@Override
-	public final Set<ITrait> getTraits() {
-	    return traits;
-	}
-	
-	@Override
 	public final <T extends ITrait> T addTrait(T trait) {
-        trait.setParent(this);
-        traits.add(trait);
-        return trait;
+//        trait.setParent(this);
+//        traits.add(trait);
+//        return trait;
+	    return TraitHandler.INSTANCE.registerTrait(this, trait); //this will be moved out externally if works, saves refactoring for testing
 	}
 	
 	@Override
 	public final <T extends ITrait> Optional<T> getTrait(Class<T> trait) {
-	    return traits.stream().filter(t -> trait.isAssignableFrom(t.getClass())).map(trait::cast).findFirst();
+	    //return traits.stream().filter(t -> trait.isAssignableFrom(t.getClass())).map(trait::cast).findFirst();
+	    return TraitHandler.INSTANCE.getTrait(this, trait);
 	}
 	
 	/**
@@ -179,8 +176,10 @@ public class CanvasObject extends Observable implements ICanvasObject {
 	@Override
 	public void setDeleted(boolean isDeleted) {
 		this.isDeleted = isDeleted;
-		this.doNotify(ObjectStatus.DELETED);
-		this.deleteObservers();
+		if (isDeleted) {
+		    this.doNotify(ObjectStatus.DELETED);
+		    this.deleteObservers();
+		}
 	}
 
 	@Override
